requirements.txt
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
defusedxml==0.7.1
executing==2.2.0
fqdn==1.5.1
icecream==2.1.4
idna==3.10
isoduration==20.11.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.2.1
pycparser==2.22
Pygments==2.19.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3987==1.3.8
rpds-py==0.24.0
six==1.17.0
tinycss2==1.4.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
uri-template==1.3.0
urllib3==2.4.0
webcolors==24.11.1
webencodings==0.5.1
